Feature - Move util and tests within protocols
