+++ protocols/tests/__init__.py
@@ -1,0 +1,7 @@
+# TODO(Greg): Refactor the tests to not use MockTestObject explicitly, we don't want to expose this
+from .generate_mock_objects import MockTestObject
+from .generate_mock_objects import get_valid_empty_cancer_participant
+from .generate_mock_objects import get_valid_empty_file
+from .generate_mock_objects import get_valid_empty_reported_somatic_variant
+from .generate_mock_objects import get_valid_report_version_control
+from .generate_mock_objects import get_valid_reported_somatic_structural_variant
