+++ protocols/tests/generate_mock_objects.py
