--- run_tests
@@ -1,1 +1,0 @@
-coverage run --source=. -m unittest discover -s tests/ && coverage report --omit="*test*" --omit="*__init__*"
