+++ run_tests.sh
@@ -1,0 +1,2 @@
+#!/usr/bin/env bash
+python -m unittest discover protocols
