+++ run_tests_with_coverage.sh
@@ -1,0 +1,2 @@
+#!/usr/bin/env bash
+coverage run --source=protocols -m unittest discover -s protocols/ && coverage report --omit="*test*" --omit="*__init__*"
